fixing integration tests
