+++ tests/extra/integration_tests/test_currency.py
@@ -1,14 +1,14 @@
 from tago import Tago
-from currency import Currency
+from tago.extra.currency import Currency
 from promise import Promise
 import os
 
 TOKEN = '37dc5a45d8e2596727edfbedc2971ebb'
 
 def test_convert():
-	result = Currency(TOKEN).current('','','')
+	result = Currency(TOKEN).convert('USD', 'GBP')
 	print result
-	if type(result) == Promise:
+	if result['from'] == 'USD':
 		assert True
 	else:
 		assert False
